Updated tag API

+++ routes/tags.js
@@ -16,6 +16,7 @@ router.put('/get', function(req, res) {
 	pg.connect(connectionString, function(err, client, done) {
 		client.query('SELECT * FROM tags WHERE post_id=\'' + req.body.id + '\';',
 		function(err, result) {
+			done();
 			if(err) {
 				console.error(err);
 				res.sendStatus(406).end();
@@ -38,7 +39,8 @@ router.put('/get/tag', function(req, res) {
 	pg.connect(connectionString, function(err, client, done) {
 		client.query('SELECT * FROM tags WHERE post_id=\'' + req.body.id + '\' AND tag=\'' + req.body.tag + '\';',
 		function(err, result) {
-			if(err) {Adds
+			done();
+			if(err) {
 				console.error(err);
 				res.sendStatus(406).end();
 			}else if(!result || result.rows.length === 0) {
@@ -55,11 +57,12 @@ router.put('/get/tag', function(req, res) {
  *
  * Adds a single tag connected to a post. Requires the poster's username and password to proceed.
  */
-router.post('/post', function(req, res) {
+router.put('/create', function(req, res) {
 	req.body = quoteFixer(req.body);
 	pg.connect(connectionString, function(err, client, done) {
 		client.query('INSERT INTO tags VALUES (\'' + req.body.id + '\', \'' + req.body.tag + '\');',
 		function(err, result) {
+			done();
 			if(err) {
 				console.error(err);
 				res.sendStatus(406).end();
@@ -78,13 +81,32 @@ router.post('/post', function(req, res) {
 router.delete('/delete', function(req, res) {
 	req.body = quoteFixer(req.body);
 	pg.connect(connectionString, function(err, client, done) {
-		client.query('DELETE FROM tags WHERE post_id=\'' + req.body.id + '\';',
-		function(err, result) {
-			if(err) {
-				console.error(err);
-				res.sendStatus(406).end();
-			}else {
-				res.sendStatus(202).end();
+		client.query('SELECT pass, salt FROM users WHERE username = \'' + req.body.username +'\';',
+		 function(err, result) {
+			 if(err) {
+				 done();
+				 console.error(err);
+				 res.sendStatus(406).end();
+			 }else if(!result || result.rows.length === 0) {
+				 done();
+				 res.sendStatus(404).end();
+			 }else {
+				 var hashpass = 'sha1$' + result.rows[0].salt + '$1$' + result.rows[0].pass;
+
+				 if(passHash.verify(req.body.pass, hashpass)) {
+					client.query('DELETE FROM tags WHERE post_id=\'' + req.body.id + '\';',
+					function(err, result) {
+						done();
+						if(err) {
+							console.error(err);
+							res.sendStatus(406).end();
+						}else {
+							res.sendStatus(201).end();
+						}
+					});
+				}else {
+					res.sendStatus(403).end();
+				}
 			}
 		});
 	});
@@ -98,15 +120,73 @@ router.delete('/delete', function(req, res) {
 router.delete('/delete/tag', function(req, res) {
 	req.body = quoteFixer(req.body);
 	pg.connect(connectionString, function(err, client, done) {
-		client.query('DELETE FROM tags WHERE post_id=\'' + req.body.id + '\' AND tag=\'' + req.body.tag + '\';',
-		function(err, result) {
-			if(err) {
-				console.error(err);
-				res.sendStatus(406).end();
-			}else {
-				res.sendStatus(202).end();
+		client.query('SELECT pass, salt FROM users WHERE username = \'' + req.body.username +'\';',
+		 function(err, result) {
+			 if(err) {
+				 done();
+				 console.error(err);
+				 res.sendStatus(406).end();
+			 }else if(!result || result.rows.length === 0) {
+				 done();
+				 res.sendStatus(404).end();
+			 }else {
+				 var hashpass = 'sha1$' + result.rows[0].salt + '$1$' + result.rows[0].pass;
+
+				 if(passHash.verify(req.body.pass, hashpass)) {
+					client.query('DELETE FROM tags WHERE post_id=\'' + req.body.id + '\' AND tag=\'' + req.body.tag + '\';',
+					function(err, result) {
+						done();
+						if(err) {
+							console.error(err);
+							res.sendStatus(406).end();
+						}else {
+							res.sendStatus(201).end();
+						}
+					});
+				}else {
+					res.sendStatus(403).end();
+				}
 			}
 		});
+	});
+});
+
+router.put('/edit/tag', function(req, res) {
+	// Nothing new to change
+  if(!req.body.new) {
+    res.sendStatus(406).end();
+  }
+
+	req.body = quoteFixer(req.body);
+	pg.connect(connectionString, function(err, client, done){
+		client.query('SELECT pass, salt FROM users WHERE username = \'' + req.body.username +'\';',
+		 function(err, result) {
+			 if(err) {
+				 done();
+				 console.error(err);
+				 res.sendStatus(406).end();
+			 }else if(!result || result.rows.length === 0) {
+				 done();
+				 res.sendStatus(404).end();
+			 }else {
+				 var hashpass = 'sha1$' + result.rows[0].salt + '$1$' + result.rows[0].pass;
+
+				 if(passHash.verify(req.body.pass, hashpass)) {
+					client.query('UPDATE FROM tags SET tag=\''+ req.body.new.tag + '\' WHERE post_id=\'' + req.body.id + '\' AND tag=\'' + req.body.tag + '\';',
+ 					function(err, result) {
+ 						done();
+ 						if(err) {
+ 							console.error(err);
+ 							res.sendStatus(406).end();
+ 						}else {
+ 							res.sendStatus(201).end();
+ 						}
+ 					});
+				 }else {
+					 res.sendStatus(403);
+				 }
+			 }
+		 });
 	});
 });
